fix: moved BertScoreWithDelimiter to qa_accuracy and updated tests
