+++ test/integration/datasets/triviaQA_sample_small.jsonl
@@ -1,8 +1,5 @@
-{"question":"Which american-born sinclair won the nobel prize for literature in 1930?","answer":"Harry Sinclair Lewis<OR>Grace Hegger<OR>Sinclair Lewis<OR>lewis harry sinclair"}
-{"question":"Where in england was dame judi dench born?","answer":"Park Grove (1895)<OR>York UA<OR>UN\/LOCODE:GBYRK<OR>York, UK<OR>City of York<OR>York, England<OR>YORK<OR>Eoferwic<OR>York"}
+{"question":"Which american-born sinclair won the nobel prize for literature in 1930?","answer":"Harry Sinclair Lewis<OR>Sinclair Lewis<OR>lewis harry sinclair"}
 {"question":"In which decade did billboard magazine first publish and american hit chart?","answer":"30's<OR>30\u2019s<OR>30s"}
 {"question":"Which city does david soul come from?","answer":"Chicago<OR>Chicago, Illinois<OR>Hog Butcher for the World<OR>Chicago, Illinois, U.S.A.<OR>The city of Chicago<OR>Chi town"}
-{"question":"Who won super bowl xx?","answer":"Chicago Bears<OR>chicago staleys<OR>chicago gators<OR>decatur staleys"}
-{"question":"Which was the first european country to abolish capital punishment?","answer":"Norv\u00e8ge<OR>Mainland Norway<OR>Norway<OR>republic of norway"}
-{"question":"What is bruce willis' real first name?","answer":"Walter<OR>walter disambiguation<OR>walter tv series"}
-{"question":"Which william wrote the novel lord of the flies?","answer":"Golding<OR>William Golding"}
+{"question":"Which was the first european country to abolish capital punishment?","answer":"Mainland Norway<OR>Norway<OR>republic of norway"}
+
